feat: Rename globals.css to globals.scss and update package.json pnpm-lock.yaml
This commit renames the file "globals.css" to "globals.scss" and updates the import statements in the file accordingly. It also adds the "scss" package as a dependency in the package.json file and updates the pnpm-lock.yaml file with the corresponding version of the "scss" package. This commit is a new feature the user and triggers a release bumping a MINOR version.