feat: Update global styles to use SASS
This commit updates the global styles in the layout.tsx file to use SASS instead of SCSS. The "globals.css" file has been renamed to "globals.scss".

The package.json file has also been updated to include the "sass" dependency with version 1.70.0. The pnpm-lock.yaml file reflects this change as well.

This commit adds a new feature for the user and will trigger a release bumping a MINOR version.