feat: Add callback route for authentication
This commit adds a new callback route for authentication. It handles the callback from the GitHub service and retrieves the authentication code. The code is then used to make a POST request to the GitHub API to get the account data. The retrieved data response- Added callback authentication of authentication code- retrieved data as JSON response